+++ lib/src/common/constants.dart
@@ -4,8 +4,7 @@ import 'package:csm_client/csm_client.dart';
 ///
 /// Stores shared key access constants for [EntityI] properties.
 final class EntityKeys {
-
-  // --> Entity properties.
+  //! [Entity] properties.
 
   /// Key for [EntityI.id].
   static const String id = 'id';
@@ -16,11 +15,24 @@ final class EntityKeys {
   /// Key for [EntityI.discriminator] property.
   static const String discriminator = 'discriminator';
 
-  // --> Named Entity properties.
+  //! [Named Entity] properties.
 
   /// Ket for [NamedEntityI.name]
   static const String name = 'name';
 
   /// Key for [NamedEntityI.description] property.
   static const String description = 'description';
+
+  //! [Referenced Entity] properties.
+
+  /// Key for [ReferencedEntityI.reference] property.
+  static const String kReference = "reference";
+
+  //! [Common Entity] properties.
+
+  /// Key for [CommonEntityI.internal] property.
+  static const String kInternal = 'internal';
+
+  /// Key for [CommonEntityI.external] property.
+  static const String kExternal = 'external';
 }